TNO-243 Fix stream bug Update data source admin ui
